port back logic for auto discovery in serverless-init
