--- cmd/serverless-sidecar/.gitignore
@@ -1,1 +1,0 @@
-serverless-sidecar
