--- cmd/serverless-sidecar/main_windows.go
@@ -1,13 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build windows
-
-//nolint:revive // TODO(SERV) Fix revive linter
-package main
-
-func main() {
-	panic("unsupported")
-}
